Changed RemoteMsg::id type from DWORD to uint32_t.

+++ src/OREMOTE2.cpp
@@ -51,11 +51,11 @@
 // ^                           ^
 // | Allocate starts here      | return variable here to the client function
 //
-RemoteMsg* Remote::new_msg(int msgId, int dataSize)
+RemoteMsg* Remote::new_msg(uint32_t msgId, int dataSize)
 {
 
-   int msgSize = sizeof(DWORD) + dataSize;
-	// <DWORD> is for the message id.
+   int msgSize = sizeof(uint32_t) + dataSize;
+	// <uint32_t> is for the message id.
 
    char* shortPtr;
 
@@ -79,7 +79,7 @@ RemoteMsg* Remote::new_msg(int msgId, int dataSize)
 
 	RemoteMsg* remoteMsgPtr = (RemoteMsg*) shortPtr;
 
-	remoteMsgPtr->id = (DWORD) msgId;
+	remoteMsgPtr->id = msgId;
 
 	return remoteMsgPtr;
 }
@@ -183,12 +183,12 @@ void Remote::send_free_msg(RemoteMsg* remoteMsgPtr, int receiverId)
 //
 // return : <char*> the pointer RemoteMsg::data_buf of the allocated message
 //
-char* Remote::new_send_queue_msg(int msgId, int dataSize)
+char* Remote::new_send_queue_msg(uint32_t msgId, int dataSize)
 {
-	char *dataPtr = send_queue[0].reserve(sizeof(short) + sizeof(DWORD) + dataSize);
+	char *dataPtr = send_queue[0].reserve(sizeof(short) + sizeof(uint32_t) + dataSize);
 
 	// ----- put the size of message ------//
-	*(short *)dataPtr = sizeof(DWORD) + dataSize;
+	*(short *)dataPtr = sizeof(uint32_t) + dataSize;
 	dataPtr += sizeof(short);
 
 	// ----- put the message id --------//
@@ -307,7 +307,7 @@ int Remote::poll_msg()
 		}
 		else
 */
-		err_when( msgListSize < sizeof(short) + sizeof(DWORD) );
+		err_when( msgListSize < sizeof(short) + sizeof(uint32_t) );
 		RemoteMsg *rMsg = (RemoteMsg *)(recvBuf + sizeof(short) );
 		int	queueToCopy = -1;
 
@@ -437,11 +437,11 @@ void Remote::process_receive_queue()
 
 				if( remoteMsgPtr->id == MSG_QUEUE_HEADER )
 				{
-					short msgNation = *(short *)(&remoteMsgPtr->data_buf[sizeof(DWORD)]);
+					short msgNation = *(short *)(&remoteMsgPtr->data_buf[sizeof(uint32_t)]);
 
 					//--- if this is the message queue of the nation we should process now ----//
 
-					if(msgNation==nationRecno )   // struct of data_buf: <DWORD> frameCount + <short> nationRecno
+					if(msgNation==nationRecno )   // struct of data_buf: <uint32_t> frameCount + <short> nationRecno
 					{
 						processFlag = 1;
 					}
@@ -504,7 +504,7 @@ void Remote::process_receive_queue()
 //
 // Pre-process a specific message type.
 //
-void Remote::process_specific_msg(DWORD msgId)
+void Remote::process_specific_msg(uint32_t msgId)
 {
 	if( !process_queue_flag )			// avoid sys.yield()
 		return;
@@ -522,11 +522,11 @@ void Remote::process_specific_msg(DWORD msgId)
 	{
 		err_when( ++loopCount > 1000 );
 		RemoteMsg* remoteMsgPtr = rqt.get_remote_msg();
-      if( remoteMsgPtr->id == msgId )
-      {
+		if( remoteMsgPtr->id == msgId )
+		{
 			remoteMsgPtr->process_msg();
-         remoteMsgPtr->id = 0;
-      }
+			remoteMsgPtr->id = 0;
+		}
 	}
 
 	enable_poll_msg();
@@ -653,13 +653,13 @@ void Remote::init_send_queue(uint32_t frameCount, short nationRecno)
 
 	// put into the queue : <message length>, MSG_QUEUE_HEADER, <frameCount>, <nationRecno>
 
-	int msgSize = sizeof(DWORD) + sizeof(uint32_t) + sizeof(short);
+	int msgSize = sizeof(uint32_t) + sizeof(uint32_t) + sizeof(short);
 	char *sendPtr = send_queue[0].reserve(sizeof(short) + msgSize);
 
 	*(short *)sendPtr = msgSize;
 	 sendPtr += sizeof(short);
-	*(DWORD *)sendPtr = MSG_QUEUE_HEADER;
-	 sendPtr += sizeof(DWORD);
+	*(uint32_t *)sendPtr = MSG_QUEUE_HEADER;
+	 sendPtr += sizeof(uint32_t);
 	*(uint32_t *)sendPtr = send_frame_count[0] = next_send_frame(nationRecno, frameCount + process_frame_delay);
 	 sendPtr += sizeof(uint32_t);
 	*(short *)sendPtr = nationRecno;
@@ -701,12 +701,12 @@ void Remote::init_receive_queue(uint32_t frameCount)
 
 			// put into the queue : <message length>, MSG_QUEUE_HEADER, <frameCount>, <nationRecno>
 
-			msgSize = sizeof(DWORD) + sizeof(uint32_t) + sizeof(short);
+			msgSize = sizeof(uint32_t) + sizeof(uint32_t) + sizeof(short);
 			receivePtr = receive_queue[n].reserve(sizeof(short) + msgSize);
 			*(short *)receivePtr = msgSize;
 			 receivePtr += sizeof(short);
-			*(DWORD *)receivePtr = MSG_QUEUE_HEADER;
-			 receivePtr += sizeof(DWORD);
+			*(uint32_t *)receivePtr = MSG_QUEUE_HEADER;
+			 receivePtr += sizeof(uint32_t);
 			*(uint32_t *)receivePtr = frameCount + n;
 			 receivePtr += sizeof(uint32_t);
 			*(short *)receivePtr = nationRecno;
@@ -714,25 +714,25 @@ void Remote::init_receive_queue(uint32_t frameCount)
 
 			// put into the queue : <message length>, MSG_NEXT_FRAME, <nationRecno>
 
-			msgSize = sizeof(DWORD) + sizeof(short);
+			msgSize = sizeof(uint32_t) + sizeof(short);
 			receivePtr = receive_queue[n].reserve(sizeof(short) + msgSize);
 
 			*(short *)receivePtr = msgSize;
 			 receivePtr += sizeof(short);
-			*(DWORD *)receivePtr = MSG_NEXT_FRAME;
-			 receivePtr += sizeof(DWORD);
+			*(uint32_t *)receivePtr = MSG_NEXT_FRAME;
+			 receivePtr += sizeof(uint32_t);
 			*(short *)receivePtr = nationRecno;
 			 receivePtr += sizeof(short);
 
 			// put into the queue : <message length>, MSG_QUEUE_TRAILER, <nationRecno>
 
-			msgSize = sizeof(DWORD) + sizeof(short);
+			msgSize = sizeof(uint32_t) + sizeof(short);
 			receivePtr = receive_queue[n].reserve(sizeof(short) + msgSize);
 
 			*(short *)receivePtr = msgSize;
 			 receivePtr += sizeof(short);
-			*(DWORD *)receivePtr = MSG_QUEUE_TRAILER;
-			 receivePtr += sizeof(DWORD);
+			*(uint32_t *)receivePtr = MSG_QUEUE_TRAILER;
+			 receivePtr += sizeof(uint32_t);
 			*(short *)receivePtr = nationRecno;
 			 receivePtr += sizeof(short);
